refactor: break down 500+ LOC handler into focused modules
- Split handler.rs into client.rs, formatting.rs, and server.rs
- Improved separation of concerns and maintainability
- Fixed test imports to use new module structure
- All tests passing (17 total)
- No functional changes, identical output behavior
