+++ src/lib.rs
@@ -1,3 +1,7 @@
-pub mod handler;
+pub mod client;
+pub mod formatting;
+pub mod server;
 
-pub use handler::*;
+pub use client::*;
+pub use formatting::*;
+pub use server::*;
